Auth backuo unit tests
